+++ playground/CHANGELOG.md
@@ -1,5 +1,110 @@
 # playground
 
+## 0.0.216-canary.16
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.16
+
+## 0.0.216-canary.15
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.15
+
+## 0.0.216-canary.14
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.14
+
+## 0.0.216-canary.13
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.13
+
+## 0.0.216-canary.12
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.12
+
+## 0.0.216-canary.11
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.11
+
+## 0.0.216-canary.10
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.10
+
+## 0.0.216-canary.9
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.9
+
+## 0.0.216-canary.8
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.8
+
+## 0.0.216-canary.7
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.7
+
+## 0.0.216-canary.6
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.6
+
+## 0.0.216-canary.5
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.5
+
+## 0.0.216-canary.4
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.4
+
+## 0.0.216-canary.3
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.3
+
+## 0.0.216-canary.2
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.3.0-canary.2
+
 ## 0.0.216-canary.1
 
 ### Patch Changes